Move coupon auto-apply hook into its own file
